2004-09-20  Tollef Fog Heen  <tfheen@raw.no>
	* loudmouth/lm-connection.c:
	  (connection_in_event): Use gsize and not gint, they have
          different sizes on 64 bit architectures.

	* loudmouth/lm-ssl.c, loudmouth/lm-internals.h:
          Adjust prototype of _lm_ssl_read to match.

+++ loudmouth/lm-connection.c
@@ -681,7 +681,7 @@ connection_in_event (GIOChannel   *source,
 		     LmConnection *connection)
 {
 	gchar     buf[IN_BUFFER_SIZE];
-	gint      bytes_read;
+	gsize      bytes_read;
 	GIOStatus status;
        
 	if (!connection->io_channel) {
